Restructurar components y estilos
